+++ lumary/db/sa/__init__.py
@@ -6,6 +6,7 @@
 from .base import Base
 from .crud import CRUDBase
 from .engine import create_db_engine, create_routing_engines
+from .mixins import AuditMixin, SoftDeleteMixin
 from .model import ModelBase
 from .session import SessionFactory, use_primary, use_replica
 
@@ -14,6 +15,8 @@ __all__ = [
     'CRUDBase',
     'create_db_engine',
     'create_routing_engines',
+    'AuditMixin',
+    'SoftDeleteMixin',
     'ModelBase',
     'SessionFactory',
     'use_primary',